+++ src/main/java/jsonvalues/spec/JsSpecs.java
@@ -1294,14 +1294,19 @@ public final class JsSpecs {
      * @throws IllegalArgumentException If the specified name already exists or the name is not valid
      */
     public static JsSpec ofNamedSpec(final String name, final JsSpec spec) {
-
-        if (spec instanceof JsObjSpec objSpec) {
-            JsSpecCache.put(name, JsObjSpecBuilder.withName(name).build(objSpec));
+        //builders already cache the specs, we need to create named specs with the builder
+        //to create a metadata object used by avro-spec library. It's more common to
+        //use the builder because there are a lot of metadata options not provided by this method
+        //(only the name). This is more commonly used by other kind of specs like oneOf(obspec1,objspec2)
+        if (requireNonNull(spec) instanceof JsObjSpec objSpec) {
+            var unused = JsObjSpecBuilder.withName(name).build(objSpec);
+            assert unused != null;
         } else if (spec instanceof JsEnum enumSpec) {
-            JsSpecCache.put(name, JsEnumBuilder.withName(name).build(enumSpec.symbols));
+            var unused = JsEnumBuilder.withName(name).build(enumSpec.symbols);
+            assert unused != null;
         } else if (spec instanceof JsFixedBinary fixed) {
-            JsSpecCache.put(name, JsFixedBuilder.withName(name).build(fixed.getSize()));
-
+            var unused = JsFixedBuilder.withName(name).build(fixed.getSize());
+            assert unused != null;
         } else JsSpecCache.put(name, spec);
         return new NamedSpec(name);
     }